pyproject.toml
[project]
name = "kayke-finance-app"
version = "0.1.0"
description = "Personal finance assistant with Telegram bot and Breno Nogueira methodology"
authors = [
    {name = "Kayke", email = "your-email@example.com"},
]
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=20.7",
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "sqlalchemy>=2.0.23",
    "psycopg2-binary>=2.9.7",
    "alembic>=1.12.1",
    "streamlit>=1.28.0",
    "plotly>=5.17.0",
    "pandas>=2.1.0",
    "groq>=0.3.0",
    "python-dotenv>=1.0.0",
    "python-dateutil>=2.8.2",
    "pydantic-settings>=2.1.0",
    "rapidfuzz>=3.14.3",
    "jinja2[fastapi]>=3.1.6",
    "requests>=2.32.5",
    "python-multipart>=0.0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kayke_finance_app"]
